refactor: rename prepare/ to prepare_retrieval/; add placeholder subpackages
Rename prepare/ → prepare_retrieval/ to better reflect the CLI command name
while still avoiding the prepare_retrieval_dataset function name collision.

Add four placeholder subpackages (README + __init__.py only, no implementation):
- preprocess/          ← medrap preprocess
- retrieve/            ← medrap retrieve
- get_embeddings/      ← medrap get-embeddings
- predict_probabilities/ ← medrap predict-probabilities
