+++ src/medrap/conf/prep/document/ordered_fields.yaml
@@ -1,4 +1,4 @@
-_target_: medrap.prepare.preparation.OrderedFieldDocumentRenderer
+_target_: medrap.prepare_retrieval.preparation.OrderedFieldDocumentRenderer
 fields: ???
 separator: "\n"
 include_field_names: false
