+++ src/medrap/get_embeddings/__init__.py
@@ -1,0 +1,1 @@
+"""Embedding inference: run a trained model over a MEDS dataset and save query or hidden-layer embeddings."""
