+++ src/medrap/predict_probabilities/__init__.py
@@ -1,0 +1,1 @@
+"""Prediction inference: run a trained model over a MEDS dataset and save predicted probabilities."""
