+++ src/medrap/preprocess/__init__.py
@@ -1,0 +1,1 @@
+"""Preprocessing of raw MEDS data into tensorized datasets for model ingestion."""
