+++ src/medrap/retrieve/__init__.py
@@ -1,0 +1,1 @@
+"""Retrieval inference: run a trained model over a MEDS dataset and save retrieved document artifacts."""
