+++ tests/test_hydra_config_integration.py
@@ -7,7 +7,7 @@ from torch import nn
 from medrap.configs import instantiate_datamodule, instantiate_model, instantiate_training_module
 from medrap.train.lightning_module import MedRAPSupervisedLightningModule
 from medrap.model.model import RetrievalAugmentedModel
-from medrap.prepare.preparation import OrderedFieldDocumentRenderer, prepare_retrieval_dataset
+from medrap.prepare_retrieval.preparation import OrderedFieldDocumentRenderer, prepare_retrieval_dataset
 from medrap.model.retrievers import HFDatasetRetriever
 from medrap.train.task import SupervisedTask
